Refactor Test Fixture 패키지 이전
기존에는 infrastructure 레이어에서 관리했는데 다른 레이어에서도 사용할 것이라 이전합니다.
