+++ ktb-be-dev-assignment/src/test/java/org/ktb/ktbbedevassignment/fixture/CompanyTestFixture.java
@@ -1,4 +1,4 @@
-package org.ktb.ktbbedevassignment.infrastructure;
+package org.ktb.ktbbedevassignment.fixture;
 
 public class CompanyTestFixture {
